add mockups for lists of books in ViewModels, replace Float by Double for rating in models

+++ app/src/main/java/com/study/mybookshelf/model/Book.kt
@@ -4,7 +4,7 @@ open class Book (
     var title: String,
     var author: String,
     var photo: Int,
-    var raiting: Float,
+    var raiting: Double,
     var isDigital: Boolean,
     var comments: String) {
 }
